implement a navigation system
