+++ android/app/src/main/java/com/coinmanager/MainActivity.java
@@ -1,6 +1,9 @@
 package com.coinmanager;
 
 import com.facebook.react.ReactActivity;
+import com.facebook.react.ReactActivityDelegate;
+import com.facebook.react.ReactRootView;
+import com.swmansion.gesturehandler.react.RNGestureHandlerEnabledRootView;
 
 public class MainActivity extends ReactActivity {
 
@@ -12,4 +15,14 @@ public class MainActivity extends ReactActivity {
     protected String getMainComponentName() {
         return "coinmanager";
     }
+
+    @Override
+    protected ReactActivityDelegate createReactActivityDelegate() {
+       return new ReactActivityDelegate(this, getMainComponentName()) {
+         @Override
+          protected ReactRootView createRootView() {
+          return new RNGestureHandlerEnabledRootView(MainActivity.this);
+     }
+   };
+ }
 }
