+++ android/settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = 'coinmanager'
+include ':react-native-gesture-handler'
+project(':react-native-gesture-handler').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-gesture-handler/android')
 
 include ':app'
