+++ package.json
@@ -10,6 +10,8 @@
   "dependencies": {
     "react": "16.8.3",
     "react-native": "0.59.9",
+    "react-native-gesture-handler": "^1.3.0",
+    "react-navigation": "^3.11.0",
     "react-redux": "^7.1.0",
     "redux": "^4.0.1",
     "redux-promise": "^0.6.0",
